pyproject.toml
[project]
name = "osrs-economy"
version = "0.1.0"
description = "An ELT project that aims to store and analyze tradable items in Old School RuneScape."
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-storage>=3.1.0",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.5",
]

[tool.ruff]
line-length = 79
